Move interfaces to root of context package.
