+++ api/src/opentrons/protocol_api/instrument_context.py
@@ -18,7 +18,7 @@ from opentrons.protocols.api_support.labware_like import LabwareLike
 from opentrons.protocol_api.module_contexts import ThermocyclerContext
 from opentrons.protocols.api_support.util import (
     FlowRates, PlungerSpeeds, Clearances, clamp_value, requires_version)
-from opentrons.protocols.context.interfaces.instrument_context import \
+from opentrons.protocols.context.instrument_context import \
     InstrumentContextInterface
 from opentrons.protocols.api_support.types import APIVersion
 from .labware import (
