--- api/src/opentrons/protocols/context/interfaces/__init__.py
@@ -1,1 +1,0 @@
-"""Interfaces of context implementations."""
