+++ api/src/opentrons/protocols/context/protocol_api/protocol_context.py
@@ -13,11 +13,11 @@ from opentrons.protocols.geometry import module_geometry
 from opentrons.protocols.geometry.deck_item import DeckItem
 from opentrons.protocols.context.protocol_api.instrument_context import \
     InstrumentContextImplementation
-from opentrons.protocols.context.interfaces.instrument_context import \
+from opentrons.protocols.context.instrument_context import \
     InstrumentContextInterface
-from opentrons.protocols.context.interfaces.labware import \
+from opentrons.protocols.context.labware import \
     LabwareInterface
-from opentrons.protocols.context.interfaces.protocol_context import \
+from opentrons.protocols.context.protocol_context import \
     ProtocolContextInterface, InstrumentDict, LoadModuleResult
 from opentrons.protocols.api_support.util import (
     AxisMaxSpeeds, HardwareToManage, HardwareManager)
